+++ client/doublezerod/internal/bgp/plugin.go
@@ -4,6 +4,7 @@ import (
 	"log/slog"
 	"net"
 	"net/netip"
+	"sync/atomic"
 	"time"
 
 	"github.com/jwhited/corebgp"
@@ -19,6 +20,10 @@ type Plugin struct {
 	RouteTable        int // kernel routing table to target for writing/removing
 	NoInstall         bool
 	RouteReaderWriter RouteReaderWriter
+
+	// These fields are used to track the initial establishment of the BGP session.
+	startedAt              time.Time
+	initialallyEstablished atomic.Bool
 }
 
 func NewBgpPlugin(
@@ -35,6 +40,7 @@ func NewBgpPlugin(
 		PeerStatusChan:    peerStatus,
 		NoInstall:         noInstall,
 		RouteReaderWriter: routeReaderWriter,
+		startedAt:         time.Now(),
 	}
 }
 
@@ -59,7 +65,17 @@ func (p *Plugin) OnOpenMessage(peer corebgp.PeerConfig, routerID netip.Addr, cap
 }
 
 func (p *Plugin) OnEstablished(peer corebgp.PeerConfig, writer corebgp.UpdateMessageWriter) corebgp.UpdateMessageHandler {
-	slog.Info("bgp: peer established")
+	if p.initialallyEstablished.CompareAndSwap(false, true) {
+		// If this is the first time we've established the session, record the duration.
+		// If the session is closed and then re-established within the lifetime of the same BGP plugin,
+		// we don't want to record the duration again since we have no starting time to compare to for
+		// those instances.
+		duration := time.Since(p.startedAt)
+		MetricSessionEstablishedDuration.WithLabelValues(peer.RemoteAddress.String()).Observe(duration.Seconds())
+		slog.Info("bgp: peer established", "duration", duration.String(), "peer", peer.RemoteAddress)
+	} else {
+		slog.Info("bgp: peer re-established", "peer", peer.RemoteAddress)
+	}
 	for _, nlri := range p.AdvertisedNLRI {
 		update, err := p.buildUpdate(nlri)
 		if err != nil {